Query updates. Files modified - [PartPSRPipe,Job1111]

+++ services/Job1111/src/com/psabackoffice/job1111/models/query/GetPsaactivityByUidCountResponse.java
@@ -18,13 +18,13 @@ public class GetPsaactivityByUidCountResponse implements Serializable {
 
     @JsonProperty("Count")
     @ColumnAlias("Count")
-    private BigInteger _count;
+    private Long _count;
 
-    public BigInteger get_count() {
+    public Long get_count() {
         return this._count;
     }
 
-    public void set_count(BigInteger _count) {
+    public void set_count(Long _count) {
         this._count = _count;
     }
